Moved customDTOs to appropriate package
